requirements.txt
pyserial~=3.4
pysms~=0.1
pytest~=5.4.3
setuptools~=41.2.0